+++ src/api/user.ts
@@ -1,6 +1,6 @@
 import ky from 'ky'
 import { AuthInput, AuthResponse } from 'types'
-import { url } from './todos'
+import { url } from './api'
 
 export const login = (json: AuthInput) =>
   ky.post(`${url}/users/login`, { json }).json<AuthResponse>()
